[성준] - 유효성 검사 추가 Feature/detail

+++ src/components/detail_page/review/ReviewBookMark.tsx
@@ -1,13 +1,21 @@
 'use client';
 import { useLoggedInUserStore } from '@/shared/store/LoggedInUser';
 import { supabase } from '@/shared/supabase/supabase';
-import React, { useState } from 'react';
+import { Database } from '@/shared/supabase/types/supabase';
+import React, { useEffect, useState } from 'react';
+
+import { CiStar } from 'react-icons/ci';
+import { FaStar } from 'react-icons/fa';
 
 function ReviewBookMark({ id }: { id: number }) {
   const [bookMarkMode, setBookMarkMode] = useState(false);
+  const [toiletLocation, setToiletLocation] = useState<Database['public']['Tables']['toilet_location']['Row']>();
   const { email: user_id } = useLoggedInUserStore((state) => state.userData);
 
   const handleAddBookMark = async () => {
+    if (!user_id) {
+      return;
+    }
     const { data, error } = await supabase
       .from('bookmark')
       .insert([{ toilet_id: id, user_id: user_id }])
@@ -19,24 +27,44 @@ function ReviewBookMark({ id }: { id: number }) {
   };
 
   const handleDeleteBookMark = async () => {
+    if (!user_id) {
+      return;
+    }
     const { error } = await supabase.from('bookmark').delete().eq(user_id, user_id);
     setBookMarkMode((prev) => !prev);
   };
+
+  useEffect(() => {
+    const getToiletLocation = async () => {
+      try {
+        const { data: toilet_location, error } = await supabase
+          .from('toilet_location')
+          .select('*')
+          .eq('toilet_id', id)
+          .single();
+        if (error) {
+          throw error.message;
+        }
+        setToiletLocation(toilet_location);
+      } catch (error) {
+        console.log(error);
+      }
+    };
+
+    getToiletLocation();
+  }, [id]);
   return (
-    <div className="absolute right-9">
-      {bookMarkMode ? (
-        <>
-          <button className="text-3xl text-amber-300" onClick={handleAddBookMark}>
-            ★
-          </button>
-        </>
-      ) : (
-        <>
-          <button className="text-3xl" onClick={handleDeleteBookMark}>
-            ☆
-          </button>
-        </>
-      )}
+    <div className="flex items-center justify-around align-middle">
+      <h2 className="text-center text-black font-medium rounded-xl text-3xl p-5 pb-6 flex gap-3">
+        {toiletLocation?.toilet_name}
+      </h2>
+      <button
+        className={`flex flex-col text-3xl ${bookMarkMode ? 'text-amber-300' : null} -translate-y-1 items-center gap-1`}
+        onClick={bookMarkMode ? handleAddBookMark : handleDeleteBookMark}
+      >
+        {bookMarkMode ? <FaStar /> : <CiStar />}
+        <p className="text-xs text-gray-400">즐겨찾기에 추가하기</p>
+      </button>
     </div>
   );
 }
